+++ unit.stop.lua
@@ -1,7 +1,7 @@
 -- Darkwinde START: unit.stop()
 
 local environmentParameter = getEnvironmentParameter(environmentID)
-if doorSwitch ~= nil and velocity_kmh <= environmentParameter["surfaceSpeedLow"] and 
+if doorSwitch ~= nil and velocity_kmh <= environmentParameter["surfaceSpeedLow"] and telemeterExists and
     round(telemeter.getDistance()) <= environmentParameter["surfaceDistanceLow"] then 
     doorSwitch.activate()
 end